+++ examples/tapo_p110.rs
@@ -2,7 +2,7 @@
 use std::{env, thread, time::Duration};
 
 use log::{info, LevelFilter};
-use tapo::{requests::EnergyDataInterval, ApiClient, P110};
+use tapo::{requests::EnergyDataInterval, ApiClient};
 use time::macros::{date, datetime};
 
 #[tokio::main]
@@ -20,7 +20,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let tapo_username = env::var("TAPO_USERNAME")?;
     let tapo_password = env::var("TAPO_PASSWORD")?;
 
-    let device = ApiClient::<P110>::new(ip_address, tapo_username, tapo_password, true).await?;
+    let device = ApiClient::new(ip_address, tapo_username, tapo_password)?
+        .p110()
+        .login()
+        .await?;
 
     info!("Turning device on...");
     device.on().await?;