+++ src/lib.rs
@@ -6,7 +6,7 @@
 //! use std::{env, thread, time::Duration};
 //!
 //! use log::{info, LevelFilter};
-//! use tapo::{requests::Color, ApiClient, L530};
+//! use tapo::{requests::Color, ApiClient};
 //!
 //! #[tokio::main]
 //! async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -23,7 +23,10 @@
 //!     let tapo_username = env::var("TAPO_USERNAME")?;
 //!     let tapo_password = env::var("TAPO_PASSWORD")?;
 //!
-//!     let device = ApiClient::<L530>::new(ip_address, tapo_username, tapo_password, true).await?;
+//!     let device = ApiClient::new(ip_address, tapo_username, tapo_password)?
+//!         .l530()
+//!         .login()
+//!         .await?;
 //!
 //!     info!("Turning device on...");
 //!     device.on().await?;
@@ -77,7 +80,6 @@
 //! See [more examples](https://github.com/mihai-dinculescu/tapo/tree/main/examples).
 
 mod api;
-mod devices;
 mod encryption;
 mod error;
 mod tapo_date_format;
@@ -86,5 +88,4 @@ pub mod requests;
 pub mod responses;
 
 pub use api::*;
-pub use devices::*;
 pub use error::*;