+++ src/responses/device_info_result/l510.rs
@@ -3,7 +3,7 @@ use serde::Deserialize;
 use crate::error::Error;
 use crate::responses::{decode_value, DefaultState, DeviceInfoResultExt, TapoResponseExt};
 
-/// Device info of [`crate::L510`]. Superset of [`crate::responses::GenericDeviceInfoResult`].
+/// Device info of Tapo L150. Superset of [`crate::responses::GenericDeviceInfoResult`].
 #[derive(Debug, Clone, Deserialize)]
 pub struct L510DeviceInfoResult {
     //