+++ src/responses/handshake_result.rs
@@ -3,7 +3,7 @@ use serde::Deserialize;
 use super::TapoResponseExt;
 
 #[derive(Debug, Deserialize)]
-pub struct HandshakeResult {
+pub(crate) struct HandshakeResult {
     pub key: String,
 }
 impl TapoResponseExt for HandshakeResult {}